tickle.h: widen struct tt_Publisher.seq_no uint16_t -> uint32_t
At the user's own explicit instruction ("seq no를 32bit로 수정하자"), fixing a real core bug
TickLE Plan found while re-sweeping the -T/-A self-throttle on real HIL (rmw_tickle/PLAN.md,
b6fb75a): pub->seq_no was uint16_t, but every downstream seq_no field it ultimately feeds -
tt_DataHeader.seq_no, tt_AckNackHeader.seq_no, tt_ReliableCacheEntry.seq_no, and pub->
peer_ack_seq_no[] - is already uint32_t. tt_Publisher_publish()'s own pub->seq_no++ silently
wrapped to 0 every 65536 publishes regardless of the wire's own 32-bit capacity - at TickLE's real
max throughput (~150-190K msg/s), reliable_throughput's own 8-second max-rate runs wrap more than
20 times per run, reusing already-delivered sequence numbers mid-stream and corrupting RELIABLE
ack tracking/DURABILITY backlog indexing for the rest of that run. Every reliable_throughput
measurement taken this session (main/v1/v2/bitmap-widening) ran under this bug.

Independently verified (not taken on the peer's own report alone): confirmed struct tt_Publisher.
seq_no's own type directly, confirmed every downstream field it feeds is uint32_t, confirmed the
exact wrap site (tt_Publisher_publish(), pub->seq_no++). Deliberately scoped to only this one
field - struct tt_Client.seq_no/struct tt_Subscriber.seq_no share the same "// transcation"
comment and field name but are unrelated: tt_Client.seq_no pairs with tt_CallRequestHeader.seq_no,
a genuinely 16-bit RPC wire field of its own; tt_Subscriber.seq_no has zero call sites anywhere in
tickle.c (unused), not implicated in this bug at all.

Not a wire format change - tt_DataHeader.seq_no/tt_AckNackHeader.seq_no were already uint32_t on
the wire; this only fixes the internal counter that feeds them, so no tt_VERSION bump is needed.
At uint32_t width the same wraparound is still structurally possible, just past ~6.3 continuous
hours at TickLE's own real max rate - not specially handled, matching this file's own existing
"an edge case far enough out not to need explicit handling yet" pragmatism elsewhere.

Verified: make test/make sanitize (ASan+UBSan)/make -C platform/freertos/clang-format --dry-run
--Werror/clang-tidy (bare invocation, 0 warnings) all clean; rmw_tickle rebuilt clean (12/12
tests) - unaffected, since it never assumes this field's own width directly.

+++ include/tickle/tickle.h
@@ -565,8 +565,23 @@ struct tt_Publisher { // extends endpoint
     struct tt_Node* node;
     struct tt_Topic* topic;
 
-    // transcation
-    uint16_t seq_no;
+    // This Publisher's own send-side sample counter (transcation) - wire seq_no is this + 1
+    // (tt_Publisher_publish()'s own data_header->seq_no = pub->seq_no + 1), starting from 1.
+    // uint32_t, matching every downstream 32-bit seq_no field that ultimately derives from it
+    // (tt_DataHeader.seq_no, tt_AckNackHeader.seq_no, tt_ReliableCacheEntry.seq_no, peer_ack_
+    // seq_no[] above) - a real bug found on real HIL (TickLE Plan, 2026-09-22): this field used
+    // to be uint16_t, silently wrapping to 0 every 65536 publishes regardless of the wire's own
+    // 32-bit capacity - at TickLE's own real max throughput (~150-190K msg/s), reliable_
+    // throughput's own 8-second max-rate runs wrap more than 20 times *per run*, reusing already-
+    // delivered sequence numbers mid-stream and corrupting RELIABLE ack tracking/DURABILITY
+    // backlog indexing for the rest of that run. At uint32_t width the same wraparound is still
+    // structurally possible, just past ~6.3 continuous hours at that same real max rate - not
+    // specially handled here, the same "an edge case far enough out not to need explicit handling
+    // yet" pragmatism this file already applies elsewhere (e.g. struct tt_Node.entity_id_base's
+    // own doc comment). Distinct from struct tt_Client.seq_no/struct tt_Subscriber.seq_no below -
+    // those pair with their own genuinely-16-bit wire counterparts (tt_CallRequestHeader.seq_no)
+    // or are unused, not affected by this same bug.
+    uint32_t seq_no;
 
     // Known Subscribers matching this Publisher's topic, learned via UPDATE announces - see
     // tt_UNICAST_PEER_THRESHOLD.
